refactor: move meaning code into meaning.rs
